requirements.txt
aiofiles==23.2.1
amqp==5.2.0
attrs==23.2.0
audioread==3.0.1
billiard==4.2.0
blinker==1.7.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.47.0
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
humanfriendly==10.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.6
imageio==2.33.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kombu==5.3.4
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
msgpack==1.0.7
networkx==3.2.1
noisereduce==3.0.0
numba==0.58.1
numpy==1.26.3
onnxruntime==1.16.3
opencv-python-headless==4.9.0.80
packaging==23.2
pillow==10.2.0
platformdirs==4.1.0
pooch==1.8.0
priority==2.0.0
prompt-toolkit==3.0.43
protobuf==4.25.1
pycparser==2.21
PyMatting==1.1.12
pyparsing==3.1.1
python-dateutil==2.8.2
Quart==0.19.4
redis==5.0.1
referencing==0.32.1
rembg==2.0.53
requests==2.31.0
rpds-py==0.16.2
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
sympy==1.12
threadpoolctl==3.2.0
tifffile==2023.12.9
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
wsproto==1.2.0
